[config] file reorganization
- move config.rs to config/mod.rs to be more rust-y
- move unit tests to inline rather than a separate file to be mroe
rust-y (plus they were really short tests)
